+++ mobipick_gui/__init__.py
@@ -1,0 +1,5 @@
+"""Mobipick Labs GUI package."""
+
+from .main_window import MainWindow, trigger_sigint
+
+__all__ = ["MainWindow", "trigger_sigint"]
